implemented client chatroom and written testcase in main.py
